requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain>=0.0.200
chromadb
sentence-transformers
transformers
python-docx
pypdf
pdfminer.six
pdf2image
pillow
pytesseract
python-magic
cryptography
python-dotenv
streamlit-authenticator
